fix: reset `framesRecorded`
Often, `framesRecorded` value was wrongly displayed on client console
due to it initialized only once when the plugin is first loaded.
This can be problematic when client is using hot reloading. It is not
reset every time `ssam:ffmpeg` message is received for each new
recording.

In the future, consider setting it to the actual frame number coming
from client.